Split off executable types to separate module
